requirements.txt
requests>=2.32
feedparser>=6.0
PyYAML>=6.0
google-generativeai>=0.8
python-dotenv>=1.0
